+++ src/Backends/MySQL/SqlObjects/MysqlPerfdata.php
@@ -77,7 +77,7 @@ class MysqlPerfdata extends MysqlModel {
             $query->bindValue($i++, $Gauge->getLabel());
             $query->bindValue($i++, ($Gauge->getTimestamp() * 1000));
             $query->bindValue($i++, $Gauge->getTimestamp());
-            $query->bindValue($i++, (double)$Gauge->getValue());
+            $query->bindValue($i++, (float)$Gauge->getValue());
             $query->bindValue($i++, $Gauge->getUnit());
         }
